Move DEFAULT_LOGGED_OUT_LABEL_PREFERENCES to const file
