+++ src/state/queries/preferences/const.ts
@@ -1,8 +1,21 @@
-import {DEFAULT_LOGGED_OUT_LABEL_PREFERENCES} from '#/state/queries/preferences/moderation'
+import {DEFAULT_LABEL_SETTINGS} from '@atproto/api'
+
 import {
   type ThreadViewPreferences,
   type UsePreferencesQueryResponse,
 } from '#/state/queries/preferences/types'
+
+/**
+ * More strict than our default settings for logged in users.
+ *
+ * Defined here rather than in `./moderation` to avoid a module-init cycle:
+ * `moderation` imports `./index`, which re-exports this file, so reading the
+ * value from `moderation` at init time lands in its temporal dead zone.
+ */
+export const DEFAULT_LOGGED_OUT_LABEL_PREFERENCES: typeof DEFAULT_LABEL_SETTINGS =
+  Object.fromEntries(
+    Object.entries(DEFAULT_LABEL_SETTINGS).map(([key, _pref]) => [key, 'hide']),
+  )
 
 export const DEFAULT_HOME_FEED_PREFS: UsePreferencesQueryResponse['feedViewPrefs'] =
   {
