+++ src/state/queries/preferences/moderation.ts
@@ -1,21 +1,9 @@
 import {useMemo} from 'react'
-import {
-  BskyAgent,
-  DEFAULT_LABEL_SETTINGS,
-  interpretLabelValueDefinitions,
-} from '@atproto/api'
+import {BskyAgent, interpretLabelValueDefinitions} from '@atproto/api'
 
 import {isNonConfigurableModerationAuthority} from '#/state/session/additional-moderation-authorities'
 import {useLabelersDetailedInfoQuery} from '../labeler'
 import {usePreferencesQuery} from './index'
-
-/**
- * More strict than our default settings for logged in users.
- */
-export const DEFAULT_LOGGED_OUT_LABEL_PREFERENCES: typeof DEFAULT_LABEL_SETTINGS =
-  Object.fromEntries(
-    Object.entries(DEFAULT_LABEL_SETTINGS).map(([key, _pref]) => [key, 'hide']),
-  )
 
 export function useMyLabelersQuery({
   excludeNonConfigurableLabelers = false,
